requirements.txt
# --- Core runtime ---
gunicorn==23.0.0
Flask==2.2.5
Werkzeug==2.2.3
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
python-dotenv==1.0.0

# --- Flask extensions ---
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
WTForms==3.1.2

# --- Database ---
SQLAlchemy==2.0.25
alembic==1.13.1
Mako==1.3.2
python-editor==1.0.4

# --- Azure ---
azure-storage-blob==12.19.1
azure-core==1.29.7
msal==1.28.0

# --- Networking ---
requests==2.31.0
urllib3==2.2.1
certifi==2024.2.2
idna==3.6
charset-normalizer==3.3.2

# --- Security / crypto ---
cryptography==42.0.2
cffi==1.16.0
pycparser==2.21
pyOpenSSL==24.0.0

# --- Utilities ---
python-dateutil==2.9.0.post0
six==1.16.0

# --- ODBC (Azure SQL) ---
pyodbc==5.2.0
